make logging more robust
