+++ src/lib/client/useLogger.ts
@@ -1,6 +1,6 @@
-/* eslint-disable no-console */
 import { useMemo } from 'react';
 import type { ILogger } from '../types';
+import { createConsoleLogger } from '../utils';
 import { WebviewLogger } from './WebviewLogger';
 import type { VsCodeApi } from './types';
 
@@ -14,18 +14,4 @@ export function useLogger(tag: string, vscode?: VsCodeApi): ILogger {
     () => (vscode === undefined ? createConsoleLogger(tag) : new WebviewLogger(vscode, tag)),
     [vscode, tag]
   );
-}
-
-export function createConsoleLogger(tag: string): ILogger {
-  return {
-    debug: (message: string, data?: Record<string, unknown>) =>
-      console.debug(`[${tag}] ${message}`, data),
-    info: (message: string, data?: Record<string, unknown>) =>
-      console.info(`[${tag}] ${message}`, data),
-    warn: (message: string, data?: Record<string, unknown>) =>
-      console.warn(`[${tag}] ${message}`, data),
-    error: (message: string, data?: Record<string, unknown>) =>
-      console.error(`[${tag}] ${message}`, data),
-    dispose: () => {},
-  };
 }
