+++ src/lib/host/logger.ts
@@ -1,6 +1,6 @@
-import * as vscode from 'vscode';
 import { LogLevel, type ILogger } from '../types';
-import { createConsoleLogger } from '../client/useLogger';
+import { createConsoleLogger } from '../utils';
+import type * as vscode from 'vscode';
 
 export const disallowedLogKeys = ['password', 'secret', 'token', 'apiKey', 'apiSecret', 'content'];
 
@@ -16,12 +16,10 @@ function removePromptsFromData<T>(dictionary: T | undefined | null): T | undefin
     return dictionary;
   }
 
-  const clone = structuredClone(dictionary) as Record<string, unknown>;
-
   try {
-    for (const key in clone) {
-      const value = clone[key];
-      if (disallowedLogKeys.includes(key)) {
+    const clone = structuredClone(dictionary) as Record<string, unknown>;
+    for (const [key, value] of Object.entries(clone)) {
+      if (disallowedLogKeys.includes(key.toLowerCase())) {
         // eslint-disable-next-line @typescript-eslint/no-dynamic-delete
         delete clone[key];
         continue;
@@ -33,12 +31,11 @@ function removePromptsFromData<T>(dictionary: T | undefined | null): T | undefin
         clone[key] = removePromptsFromData(value as Record<string, unknown>) as unknown;
       }
     }
+    return clone as unknown as T;
   } catch (error) {
     console.error('Error processing log data:', error);
     return {} as T;
   }
-
-  return clone as unknown as T;
 }
 
 /**
@@ -48,7 +45,7 @@ function removePromptsFromData<T>(dictionary: T | undefined | null): T | undefin
 class LoggerImpl {
   // eslint-disable-next-line sonarjs/public-static-readonly
   static outputChannel: vscode.OutputChannel | undefined = undefined;
-  private static readonly consoleLogger = createConsoleLogger('RVW-IPC');
+  private static readonly consoleLogger = createConsoleLogger('RVW');
 
   public static debug(message: string, data: Record<string, unknown> | undefined = undefined) {
     this.log(LogLevel.DEBUG, message, data);
@@ -68,33 +65,27 @@ class LoggerImpl {
 
   public static dispose() {
     this.outputChannel?.dispose();
+    this.outputChannel = undefined;
   }
 
   private static log(level: LogLevel, message: string, data: Record<string, unknown> | undefined) {
     const timestamp = new Date().toISOString().split('T')[1];
-    const levelStr = LogLevel[level] || 'UNKNOWN';
     const cleanedData = removePromptsFromData(data);
-    const line = `[${timestamp}] [${levelStr}] ${message}`;
     if (this.outputChannel === undefined) {
-      switch (level) {
-        case LogLevel.DEBUG: {
-          this.consoleLogger.debug(line, cleanedData);
-          break;
-        }
-        case LogLevel.INFO: {
-          this.consoleLogger.info(line, cleanedData);
-          break;
-        }
-        case LogLevel.WARN: {
-          this.consoleLogger.warn(line, cleanedData);
-          break;
-        }
-        case LogLevel.ERROR: {
-          this.consoleLogger.error(line, cleanedData);
-          break;
+      const methodName = LogLevel[level].toLowerCase() as
+        | undefined
+        | keyof Omit<ILogger, 'dispose'>;
+      if (methodName !== undefined && typeof this.consoleLogger[methodName] === 'function') {
+        if (cleanedData === undefined) {
+          this.consoleLogger[methodName](`[${timestamp}] ${message}`);
+        } else {
+          this.consoleLogger[methodName](`[${timestamp}] ${message}`, cleanedData);
         }
       }
     } else {
+      const levelStr = LogLevel[level] || 'UNKNOWN';
+      const line = `[${timestamp}] [${levelStr}] ${message}`;
+
       if (cleanedData === undefined) {
         this.outputChannel.appendLine(line);
       } else {
@@ -108,9 +99,17 @@ class LoggerImpl {
   }
 }
 
-export const Logger: ILogger & { setOutputChannel: (outputChannel: vscode.OutputChannel) => void } =
+export const Logger: ILogger & {
+  setOutputChannel: (outputChannel: vscode.OutputChannel | undefined) => void;
+} =
+  /**
+   * Sets the VS Code output channel for the logger.
+   * The logger takes ownership of the channel and will dispose it
+   * when `Logger.dispose()` is called or when a new channel is set.
+   * @param outputChannel The output channel to use for logging.
+   */
   {
-    setOutputChannel: (outputChannel: vscode.OutputChannel) => {
+    setOutputChannel: (outputChannel: vscode.OutputChannel | undefined) => {
       LoggerImpl.dispose();
       LoggerImpl.outputChannel = outputChannel;
     },
